update min max data type

+++ src/main/java/springboot/demo/mysql/vo/specs/TMDoubleSpecs.java
@@ -13,7 +13,7 @@ import springboot.demo.mysql.vo.ThingModelSpecs;
 public class TMDoubleSpecs extends ThingModelSpecs {
     Integer efficientNum;
 
-    Integer min;
+    Double min;
 
-    Integer max;
+    Double max;
 }
